ATR-855: reworked event handler collection to work in a natural order
