+++ src/Acuminator/Acuminator.Utilities/Roslyn/Semantic/SemanticModel/PXGraph/PXGraphWithEvents/PXGraphEventSemanticModel.cs
@@ -155,7 +155,7 @@ namespace Acuminator.Utilities.Roslyn.Semantic.PXGraph
 			_cancellation = cancellation;
 			BaseGraphModel = baseGraphModel;
 
-			var eventsCollector = new EventHandlersCollector(this, PXContext);
+			var eventsCollector = new NaturalOrderEventHandlersCollector(this, PXContext);
 			eventsCollector.CollectGraphEventHandlers(_cancellation);
 
 			OrderedEventHandlers  = GraphEventHandlersStorage.FromCollected(eventsCollector);
